Vérification du token pour chaque reqête
+ CommandeController ajout de getCommandeByClientId

+++ CircuitsCourtsApi/src/main/java/com/lifat/CircuitsCourtsApi/controller/ProduitController.java
@@ -1,10 +1,12 @@
 package com.lifat.CircuitsCourtsApi.controller;
 
+import com.lifat.CircuitsCourtsApi.JWTEndPointsProtection.JwtUtil;
 import com.lifat.CircuitsCourtsApi.model.Produit;
 import com.lifat.CircuitsCourtsApi.service.ProduitService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
+@CrossOrigin(origins = "http://localhost:3000")
 @RestController
 @RequestMapping("/api")
 public class ProduitController {
@@ -12,19 +14,29 @@ public class ProduitController {
     @Autowired
     private ProduitService produitService;
 
+    @Autowired
+    private JwtUtil jwtUtil;
+
     @GetMapping("/produits")
-    public Iterable<Produit> getProduits() {
-        return produitService.getProduits();
+    public ResponseEntity<?> getProduits(@RequestParam String key) throws Exception {
+        if(jwtUtil.isValidToken(key)){
+            return ResponseEntity.ok(produitService.getProduits()) ;
+        }else return ResponseEntity.badRequest().body("Invalid Token");
     }
 
     @PostMapping("/produits")
-    public Produit saveProduit(@RequestBody Produit produit) {
-        return produitService.saveProduit(produit);
+    public ResponseEntity<?> saveProduit(@RequestBody Produit produit, @RequestParam String key) throws Exception {
+        if(jwtUtil.isValidToken(key)){
+            return ResponseEntity.ok(produitService.saveProduit(produit));
+        }else return ResponseEntity.badRequest().body("Invalid Token");
     }
 
     @DeleteMapping("/produits/{id}")
-    public void deleteProduit(@PathVariable Long id) {
-        produitService.deleteProduit(id);
+    public ResponseEntity<?> deleteProduit(@PathVariable Long id, @RequestParam String key) throws Exception {
+        if(jwtUtil.isValidToken(key)){
+            produitService.deleteProduit(id);
+            return ResponseEntity.ok("Produit with id " + id + " has been deleted.");
+        }else return ResponseEntity.badRequest().body("Invalid Token");
     }
 
 }
